Fix Menu cb throwing on MenuAction_Cancel
Fix a bug where a MenuHandler callback could throw if the Menu got
cancelled. This bug resulted from the incorrect assumption that a
MenuHandler invoked with MENU_TIME_FOREVER could never cancel, which is
untrue since other menus can override what's currently being shown to
client.

While this fixes the bug, the underlying issue of an external menu
stepping over the veto menu remains. It might be worth adding some kind
of loop that periodically re-displays any pending veto menus, but that's
outside the scope of this hotfix.